Adding ISO8601 duration support to Azure.CSharp generator
This is a fix for this issue: https://github.com/Azure/autorest/issues/390
Some of the templates are duplicated between the CSharp and Azure.CSharp
flavors, so we have to be careful to make changes in both places. To validate
the fix, I simply copied some of the duration tests to the Azure.CSharp test
project. This was for expedience. We should consider better approaches to
ensuring parity of test coverage between flavors.