Relax version constrain on JSON gem, to match version in Windows release
The JSON gem carries with it native C extensions, that require a
compiler toolchain be present on any machine that installs the latest
1.8.3 gem (since there are no precompiled binaries available for this
gem).

Ruby carries the following JSON gem versions with the distribution by
default:

* JSON 1.4.2 with 1.9.2 - https://github.com/ruby/ruby/blob/ruby_1_9_2/ext/json/lib/json/version.rb
* JSON 1.5.5 with 1.9.3 - https://github.com/ruby/ruby/blob/ruby_1_9_3/ext/json/lib/json/version.rb
* JSON 1.7.7 with 2.0.0 series - https://github.com/ruby/ruby/blob/ruby_2_0_0/ext/json/lib/json/version.rb
* JSON 1.8.1 with 2.1.0 - through 2.1.7 - https://github.com/ruby/ruby/blob/v2_1_7/ext/json/lib/json/version.rb
* JSON 1.8.1 with 2.2.0 - through 2.2.3 - https://github.com/ruby/ruby/blob/v2_2_3/ext/json/lib/json/version.rb

By relaxing the constraint we allow people to use the SDK without
installing DevKit (and the compiler toolchain) everywhere.