GetDerivedType should consider only the derived types of the provided base type (#1352)
* Update PolymorphicJsonConverter.cs

* Added PolymorphicDeserializationConsidersOnlyOwnDerivedTypes  test

This is part of the fix for the GetDerivedTypes bug where the method was
considering all the types in the Assembly for a potential derived type
(based on the discriminator value). This search has been scoped to just
the types inheriting from the base type, as part of this fix.