[Go] Calculating Version & UserAgent at Gen time.
The previous update to Version and UserAgent created some benign race-conditions. Despite the fact that there was no behavioral difference, it still triggered Go errors when "-race" was used. To fix this, we're offloading the calculation of Version and UserAgent information into the Generator. Note, this PR doesn't attempt to fix the broader issue of the fact that UserAgents are always tied to Azure-SDK-for-Go and ARM.