Removing APIVersion from client.
Instead, a const is generated in each method that will serve to provide the APIVersion if present. This was done in order to ensure that composite swaggers which have multiple APIVersions associated with their operations are supported.

One upside, is that APIVersion will no longer be present in swaggers that do not contain it. Pet Store should not have a reference to APIVersion.

Regenerating samples.