Fix documented parent for x-ms-external
The x-ms-external extension is only accepted on Schema Objects, not
"Definitions Object" (as currently linked) or "Definition Objects" (as
currently stated - not specified by OpenAPI).

Note: I confirmed (using C#) that the parent of x-ms-external is not
restricted to Schema Objects which are children of Definitions Object.
It worked on every Schema Object tested, including property schemas
with class names generated from property names.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>